refactor(GhostView): 使用`Main.DelayedProcesses`计数死亡时间，减少开销

+++ src/GhostView/Strategies/RespawnTimeSelector.cs
@@ -4,11 +4,11 @@ namespace GhostView.Strategies;
 
 public class RespawnTimeSelector
 {
-    private readonly double _normalSeconds = TShock.Config.Settings.RespawnSeconds;
+    private readonly int _normalSeconds = TShock.Config.Settings.RespawnSeconds;
 
-    private readonly double _bossSeconds = TShock.Config.Settings.RespawnBossSeconds;
+    private readonly int _bossSeconds = TShock.Config.Settings.RespawnBossSeconds;
 
-    public double GetRespawnSeconds(bool isBossAlive)
+    public int GetRespawnSeconds(bool isBossAlive)
     {
         return isBossAlive ? this._bossSeconds : this._normalSeconds;
     }
